requirements.txt
PyMuPDF>=1.23.0
python-docx>=0.8.11
Pillow>=9.0.0
openpyxl>=3.0.0
pyinstaller>=5.0.0 